+++ src/main/java/org/opentripplanner/routing/algorithm/raptor/transit/request/RaptorRoutingRequestTransitData.java
@@ -6,6 +6,7 @@ import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
+import org.opentripplanner.routing.algorithm.raptor.transit.RaptorTransferIndex;
 import java.util.function.IntFunction;
 import javax.annotation.Nullable;
 import org.opentripplanner.model.transfer.TransferService;
@@ -43,7 +44,7 @@ public class RaptorRoutingRequestTransitData implements RaptorTransitDataProvide
   /**
    * Transfers by stop index
    */
-  private final List<List<RaptorTransfer>> transfers;
+  private final RaptorTransferIndex transfers;
 
   private final ZonedDateTime startOfTime;
 
@@ -80,17 +81,16 @@ public class RaptorRoutingRequestTransitData implements RaptorTransitDataProvide
     );
   }
 
-  /**
-   * Gets all the transfers starting at a given stop
-   */
   @Override
-  public Iterator<RaptorTransfer> getTransfers(int stopIndex) {
-    return transfers.get(stopIndex).iterator();
+  public Iterator<RaptorTransfer> getTransfersFromStop(int stopIndex) {
+    return transfers.getForwardTransfers().get(stopIndex).iterator();
   }
 
-  /**
-   * Gets all the unique trip patterns touching a set of stops
-   */
+  @Override
+  public Iterator<? extends RaptorTransfer> getTransfersToStop(int stopIndex) {
+    return transfers.getReversedTransfers().get(stopIndex).iterator();
+  }
+
   @Override
   public Iterator<? extends RaptorRoute<TripSchedule>> routeIterator(IntIterator stops) {
     Set<RaptorRoute<TripSchedule>> activeTripPatternsForGivenStops = new HashSet<>();